+++ src/main/java/com/ajromero/store/IntArray.java
@@ -3,7 +3,7 @@ package com.ajromero.store;
 public class IntArray extends StoreOperation {
   private int[] inputArray;
 
-  public IntArray(int[] inputArray, int min, int max, float avg) {
+  public IntArray(int[] inputArray, int min, int max, double avg) {
     super(min, max, avg);
     this.inputArray = inputArray;
   }
